Patch 0.1.3: Refactored Project page -> grid -> components (+ added animation to accordion)
