+++ src/pages/blocks/ProfessionGrid.js
@@ -1,8 +1,8 @@
 import Grid from "@material-ui/core/Grid";
 import Typography from "@material-ui/core/Typography";
 import Divider from "@material-ui/core/Divider";
-import SpringImageCard from "../../components/SpringImageCard";
-import ImageCard from "../../components/ImageCard";
+import SpringImageCard from "../../components/subcomponents/SpringImageCard";
+import ImageCard from "../../components/subcomponents/ImageCard";
 import PdfViewer from "../../components/subcomponents/PdfViewer";
 import React from "react";
 import {makeStyles} from "@material-ui/core/styles";
